+++ src/test/kotlin/db/tables/GroupChatUsersUtil.kt
@@ -8,7 +8,4 @@ fun GroupChatUsers.read(): List<Int> = transaction {
     selectAll().orderBy(GroupChatUsers.id).map { it[GroupChatUsers.id].value }
 }
 
-fun GroupChatUsers.canUsersLeave(chatId: Int, vararg userIdList: Int): Boolean =
-    canUsersLeave(chatId, userIdList.toList())
-
 fun GroupChatUsers.makeAdmins(chatId: Int, vararg userIdList: Int): Unit = makeAdmins(chatId, userIdList.toList())